+++ src/validators/commands/trading-wallet/WithdrawCommandValidator.js
@@ -5,7 +5,8 @@ const withdrawSchema = customJoiValidator.object()
   .keys({
     from: customJoiValidator.address().ethereum().required(),
     to: customJoiValidator.address().ethereum().required(),
-    privateKeyFilePath: customJoiValidator.path().existFile(),
+    keystoreFilePath: customJoiValidator.path().existFile().required(),
+    keystorePassword: customJoiValidator.string().required(),
     quantity: customJoiValidator.bigNumber().valid().required(),
     token: customJoiValidator.address().ethereum().required(),
     draft: customJoiValidator.boolean(),